+++ Estrapy/autoupdate.py
@@ -1,5 +1,5 @@
-import requests
 import os
+from .http import Requester
 from packaging import version
 from .__init__ import __version__
 
@@ -8,10 +8,10 @@ __all__ = ("AutoUpdate",)
 
 class AutoUpdate:
     def __init__(self):
-        self.url = "https://estra-api.vercel.app/version/"
+        self.requester = Requester(custom_url="https://estra-api.vercel.app/version/")
 
     def update(self):
-        req = requests.post(url=self.url, json={"language": "python"}).json()
+        req = self.requester.post_api(json={"language": "python"})
 
         if version.parse(__version__) == version.parse(req.get("version")):
             return
@@ -20,7 +20,7 @@ class AutoUpdate:
             os.system("pip install Estrapy-API --upgrade")
 
     def reminder(self):
-        req = requests.post(url=self.url, json={"language": "python"}).json()
+        req = self.requester.post_api(json={"language": "python"})
 
         if version.parse(__version__) == version.parse(req.get("version")):
             return