+++ Estrapy/property.py
@@ -6,8 +6,9 @@ __all__ = (
 
 
 class PropertiesManager(object):
-    def __init__(self, *args, **kwargs) -> None:
-        self.url = kwargs.get("url")
+    def __init__(self, **kwargs) -> None:
+        self.url = kwargs.get("url")  # One request
+        self.urls = kwargs.get("urls")  # Multiple requests
         self.text = kwargs.get("text")
         self.type = kwargs.get("type")
         self.player = kwargs.get("player")
@@ -15,21 +16,30 @@ class PropertiesManager(object):
         self.percentage = kwargs.get("percentage")
         self.total = kwargs.get("total")
         self.with_account = kwargs.get("with_account")
+        self.original_response = kwargs.get("original_response")
 
     @property
     def url(self) -> str:
         return self._url
 
     @url.setter
-    def url(self, value):
+    def url(self, value) -> None:
         self._url = value
+
+    @property
+    def urls(self) -> list[str]:
+        return self._urls
+
+    @urls.setter
+    def urls(self, value) -> None:
+        self._urls = value
 
     @property
     def text(self) -> str:
         return self._text
 
     @text.setter
-    def text(self, value):
+    def text(self, value) -> None:
         self._text = value
 
     @property
@@ -37,7 +47,7 @@ class PropertiesManager(object):
         return self._type
 
     @type.setter
-    def type(self, value):
+    def type(self, value) -> None:
         self._type = value
 
     @property
@@ -45,7 +55,7 @@ class PropertiesManager(object):
         return self._player
 
     @type.setter
-    def player(self, value):
+    def player(self, value) -> None:
         self._player = value
 
     @property
@@ -53,7 +63,7 @@ class PropertiesManager(object):
         return self._character_name
 
     @character_name.setter
-    def character_name(self, value):
+    def character_name(self, value) -> None:
         self._character_name = value
 
     @property
@@ -61,7 +71,7 @@ class PropertiesManager(object):
         return self._percentage
 
     @percentage.setter
-    def percentage(self, value):
+    def percentage(self, value) -> None:
         self._percentage = value
 
     @property
@@ -69,7 +79,7 @@ class PropertiesManager(object):
         return self._total
 
     @total.setter
-    def total(self, value):
+    def total(self, value) -> None:
         self._total = value
 
     @property
@@ -77,8 +87,114 @@ class PropertiesManager(object):
         return self._with_account
 
     @with_account.setter
-    def with_account(self, value):
+    def with_account(self, value) -> None:
         self._with_account = value
+
+    @property
+    def original_response(self) -> dict:
+        return self._original_response
+
+    @original_response.setter
+    def original_response(self, value) -> None:
+        self._original_response = value
+
+
+class AccountProperties(object):
+    def __init__(self, **kwargs) -> None:
+        self.username = kwargs.get("username")
+        self.user_id = kwargs.get("user_id")
+        self.client_id = kwargs.get("client_id")
+        self.client_secret = kwargs.get("client_secret")
+
+    @property
+    def username(self) -> str:
+        return self._username
+
+    @username.setter
+    def username(self, value) -> None:
+        self._username = value
+
+    @property
+    def user_id(self) -> str:
+        return self._user_id
+
+    @user_id.setter
+    def user_id(self, value) -> None:
+        self._user_id = value
+
+    @property
+    def client_id(self) -> str:
+        return self._client_id
+
+    @client_id.setter
+    def client_id(self, value) -> None:
+        self._client_id = value
+
+    @property
+    def client_secret(self) -> str:
+        return self._client_secret
+
+    @client_secret.setter
+    def client_secret(self, value) -> None:
+        self._client_secret = value
+
+
+class AccountStatistics(object):
+    def __init__(self, **kwargs):
+        self.sfw = kwargs.get("sfw")
+        self.nsfw = kwargs.get("nsfw")
+        self.games = kwargs.get("games")
+        self.anigames = kwargs.get("anigames")
+        self.osu = kwargs.get("osu")
+        self.total_requests_user = kwargs.get("total_requests_user")
+
+    @property
+    def sfw(self) -> dict:
+        return self._sfw
+
+    @sfw.setter
+    def sfw(self, value) -> None:
+        self._sfw = value
+
+    @property
+    def nsfw(self) -> dict:
+        return self._nsfw
+
+    @nsfw.setter
+    def nsfw(self, value) -> None:
+        self._nsfw = value
+
+    @property
+    def games(self) -> dict:
+        return self._games
+
+    @games.setter
+    def games(self, value) -> None:
+        self._games = value
+
+    @property
+    def anigames(self) -> dict:
+        return self._anigames
+
+    @anigames.setter
+    def anigames(self, value) -> None:
+        self._anigames = value
+
+    @property
+    def osu(self) -> dict:
+        return self._osu
+
+    @osu.setter
+    def osu(self, value) -> None:
+        self._osu = value
+
+    @property
+    def total_requests_user(self) -> int:
+        return self._total_requests_user
+
+    @total_requests_user.setter
+    def total_requests_user(self, value) -> None:
+        self._total_requests_user = value
 
 
 class OsuProfileProperties(object):
@@ -129,9 +245,11 @@ class OsuProfileProperties(object):
         "loved_beatmapset_count",
         "mapping_follower_count",
         "monthly_playcounts",
+        "nominated_beatmapset_count",
         "page",
         "pending_beatmapset_count",
         "previous_usernames",
+        "rank_highest",
         "ranked_beatmapset_count",
         "replays_watched_counts",
         "scores_best_count",
@@ -147,7 +265,7 @@ class OsuProfileProperties(object):
         "unranked_beatmapset_count",
     )
 
-    def __init__(self, _json: dict = None):
+    def __init__(self, _json: dict = None) -> None:
         if _json:
             for key, value in _json.items():
                 setattr(self, key, value)
@@ -187,7 +305,7 @@ class OsuBeatmapProperties(object):
         "max_combo",
     )
 
-    def __init__(self, _json: dict = None):
+    def __init__(self, _json: dict = None) -> None:
         if _json:
             for key, value in _json.items():
                 setattr(self, key, value)