+++ Estrapy/sfw.py
@@ -1,4 +1,4 @@
-from .http import get_api, post_api
+from .http import Requester
 from .base import Base
 from .property import PropertiesManager
 from typing import Optional
@@ -7,11 +7,15 @@ __all__ = ("Sfw",)
 
 
 class Sfw:
-    __slots__ = ("token_user", "user_id")
+    __slots__ = ("client_id", "client_secret", "requester", "base")
 
-    def __init__(self, token_user: Optional[str] = None, user_id: Optional[int] = None):
-        self.token_user = token_user
-        self.user_id = user_id
+    def __init__(
+        self, client_id: Optional[str] = None, client_secret: Optional[int] = None
+    ):
+        self.client_id = client_id
+        self.client_secret = client_secret
+        self.requester = Requester()
+        self.base = Base()
 
     async def run(self, generate: Optional[int] = None) -> PropertiesManager:
         """
@@ -39,24 +43,26 @@ class Sfw:
         """
 
         route = "sfw/run"
-        output = get_api(route=route)
+        output = self.requester.get_api(route=route)
 
-        if self.token_user and self.user_id:
+        if self.client_id and self.client_secret:
             _json = {
-                "token_user": self.token_user,
-                "user_id": self.user_id,
+                "client_id": self.client_id,
+                "client_secret": self.client_secret,
             }
-            output = post_api(route=route, json=_json)
+            output = self.requester.post_api(route=route, json=_json)
 
         properties = PropertiesManager(
             url=output.get("link"),
             type=output.get("type"),
             total=output.get("total_image"),
             with_account=output.get("with_account"),
+            original_response=output,
         )
 
         if generate:
-            return await Base.produce(total=generate, route=route, type="link")
+            output = await self.base.produce(total=generate, route=route, type="link")
+            properties = PropertiesManager(urls=output)
 
         return properties
 
@@ -86,24 +92,26 @@ class Sfw:
         """
 
         route = "sfw/hug"
-        output = get_api(route=route)
+        output = self.requester.get_api(route=route)
 
-        if self.token_user and self.user_id:
+        if self.client_id and self.client_secret:
             _json = {
-                "token_user": self.token_user,
-                "user_id": self.user_id,
+                "client_id": self.client_id,
+                "client_secret": self.client_secret,
             }
-            output = post_api(route=route, json=_json)
+            output = self.requester.post_api(route=route, json=_json)
 
         properties = PropertiesManager(
             url=output.get("link"),
             type=output.get("type"),
             total=output.get("total_image"),
             with_account=output.get("with_account"),
+            original_response=output,
         )
 
         if generate:
-            return await Base.produce(total=generate, route=route, type="link")
+            output = await self.base.produce(total=generate, route=route, type="link")
+            properties = PropertiesManager(urls=output)
 
         return properties
 
@@ -133,24 +141,26 @@ class Sfw:
         """
 
         route = "sfw/smile"
-        output = get_api(route=route)
+        output = self.requester.get_api(route=route)
 
-        if self.token_user and self.user_id:
+        if self.client_id and self.client_secret:
             _json = {
-                "token_user": self.token_user,
-                "user_id": self.user_id,
+                "client_id": self.client_id,
+                "client_secret": self.client_secret,
             }
-            output = post_api(route=route, json=_json)
+            output = self.requester.post_api(route=route, json=_json)
 
         properties = PropertiesManager(
             url=output.get("link"),
             type=output.get("type"),
             total=output.get("total_image"),
             with_account=output.get("with_account"),
+            original_response=output,
         )
 
         if generate:
-            return await Base.produce(total=generate, route=route, type="link")
+            output = await self.base.produce(total=generate, route=route, type="link")
+            properties = PropertiesManager(urls=output)
 
         return properties
 
@@ -180,24 +190,26 @@ class Sfw:
         """
 
         route = "sfw/headpat"
-        output = get_api(route=route)
+        output = self.requester.get_api(route=route)
 
-        if self.token_user and self.user_id:
+        if self.client_id and self.client_secret:
             _json = {
-                "token_user": self.token_user,
-                "user_id": self.user_id,
+                "client_id": self.client_id,
+                "client_secret": self.client_secret,
             }
-            output = post_api(route=route, json=_json)
+            output = self.requester.post_api(route=route, json=_json)
 
         properties = PropertiesManager(
             url=output.get("link"),
             type=output.get("type"),
             total=output.get("total_image"),
             with_account=output.get("with_account"),
+            original_response=output,
         )
 
         if generate:
-            return await Base.produce(total=generate, route=route, type="link")
+            output = await self.base.produce(total=generate, route=route, type="link")
+            properties = PropertiesManager(urls=output)
 
         return properties
 
@@ -227,24 +239,26 @@ class Sfw:
         """
 
         route = "sfw/poke"
-        output = get_api(route=route)
+        output = self.requester.get_api(route=route)
 
-        if self.token_user and self.user_id:
+        if self.client_id and self.client_secret:
             _json = {
-                "token_user": self.token_user,
-                "user_id": self.user_id,
+                "client_id": self.client_id,
+                "client_secret": self.client_secret,
             }
-            output = post_api(route=route, json=_json)
+            output = self.requester.post_api(route=route, json=_json)
 
         properties = PropertiesManager(
             url=output.get("link"),
             type=output.get("type"),
             total=output.get("total_image"),
             with_account=output.get("with_account"),
+            original_response=output,
         )
 
         if generate:
-            return await Base.produce(total=generate, route=route, type="link")
+            output = await self.base.produce(total=generate, route=route, type="link")
+            properties = PropertiesManager(urls=output)
 
         return properties
 
@@ -274,24 +288,26 @@ class Sfw:
         """
 
         route = "sfw/bite"
-        output = get_api(route=route)
+        output = self.requester.get_api(route=route)
 
-        if self.token_user and self.user_id:
+        if self.client_id and self.client_secret:
             _json = {
-                "token_user": self.token_user,
-                "user_id": self.user_id,
+                "client_id": self.client_id,
+                "client_secret": self.client_secret,
             }
-            output = post_api(route=route, json=_json)
+            output = self.requester.post_api(route=route, json=_json)
 
         properties = PropertiesManager(
             url=output.get("link"),
             type=output.get("type"),
             total=output.get("total_image"),
             with_account=output.get("with_account"),
+            original_response=output,
         )
 
         if generate:
-            return await Base.produce(total=generate, route=route, type="link")
+            output = await self.base.produce(total=generate, route=route, type="link")
+            properties = PropertiesManager(urls=output)
 
         return properties
 
@@ -321,24 +337,26 @@ class Sfw:
         """
 
         route = "sfw/neko"
-        output = get_api(route=route)
+        output = self.requester.get_api(route=route)
 
-        if self.token_user and self.user_id:
+        if self.client_id and self.client_secret:
             _json = {
-                "token_user": self.token_user,
-                "user_id": self.user_id,
+                "client_id": self.client_id,
+                "client_secret": self.client_secret,
             }
-            output = post_api(route=route, json=_json)
+            output = self.requester.post_api(route=route, json=_json)
 
         properties = PropertiesManager(
             url=output.get("link"),
             type=output.get("type"),
             total=output.get("total_image"),
             with_account=output.get("with_account"),
+            original_response=output,
         )
 
         if generate:
-            return await Base.produce(total=generate, route=route, type="link")
+            output = await self.base.produce(total=generate, route=route, type="link")
+            properties = PropertiesManager(urls=output)
 
         return properties
 
@@ -368,24 +386,26 @@ class Sfw:
         """
 
         route = "sfw/highfive"
-        output = get_api(route=route)
+        output = self.requester.get_api(route=route)
 
-        if self.token_user and self.user_id:
+        if self.client_id and self.client_secret:
             _json = {
-                "token_user": self.token_user,
-                "user_id": self.user_id,
+                "client_id": self.client_id,
+                "client_secret": self.client_secret,
             }
-            output = post_api(route=route, json=_json)
+            output = self.requester.post_api(route=route, json=_json)
 
         properties = PropertiesManager(
             url=output.get("link"),
             type=output.get("type"),
             total=output.get("total_image"),
             with_account=output.get("with_account"),
+            original_response=output,
         )
 
         if generate:
-            return await Base.produce(total=generate, route=route, type="link")
+            output = await self.base.produce(total=generate, route=route, type="link")
+            properties = PropertiesManager(urls=output)
 
         return properties
 
@@ -415,23 +435,25 @@ class Sfw:
         """
 
         route = "sfw/slap"
-        output = get_api(route=route)
+        output = self.requester.get_api(route=route)
 
-        if self.token_user and self.user_id:
+        if self.client_id and self.client_secret:
             _json = {
-                "token_user": self.token_user,
-                "user_id": self.user_id,
+                "client_id": self.client_id,
+                "client_secret": self.client_secret,
             }
-            output = post_api(route=route, json=_json)
+            output = self.requester.post_api(route=route, json=_json)
 
         properties = PropertiesManager(
             url=output.get("link"),
             type=output.get("type"),
             total=output.get("total_image"),
             with_account=output.get("with_account"),
+            original_response=output,
         )
 
         if generate:
-            return await Base.produce(total=generate, route=route, type="link")
+            output = await self.base.produce(total=generate, route=route, type="link")
+            properties = PropertiesManager(urls=output)
 
         return properties